Custom default theme better integrated with current middlebury theme
Made some CSS changes and changes to the display action that makes
the pages. Added a middlebury logo at the top that links back to the
main middlebury page (common navigation element), as well as positioning,
color, and typography changes.
Styled the login button to appear inside the main "background" and to
stand out more.
Removed the languages select, didn't seem to do anything.