Added conversion status feedback to the user-interface.
Shel was asking about feedback during the conversion process and this
is an attempt to solve the issue he raised. We don't have an easy way
to know what percentage the process has completed, but at least now
we can see how many videos are in line before ours and how long we
have been encoding for.