Fix DuplicateKeyException when uploading an incompletely deleted file. WHD#509223
In cases where an upload failed early or a broken file was deleted,
it is possible that an entry in the middmedia_metadata table
already exists for the file even though the file itself is gone.

This change allows that metadata entry to be updated if it already
exists instead of breaking the upload process.